[세부 일정 설정 화면] 마커 색상 및 날짜 변경 오류 해결

+++ app/src/main/java/com/thequietz/travelog/schedule/repository/ScheduleRepository.kt
@@ -47,9 +47,11 @@ class ScheduleRepository @Inject constructor(
         scheduleDetails: List<ScheduleDetailModel>
     ) {
         coroutineScope.launch {
-            scheduleDetailDao.deleteScheduleDetailsByScheduleId(scheduleDetails[0].scheduleId)
-            scheduleDetails.forEach {
-                scheduleDetailDao.insert(it)
+            if (scheduleDetails.isNotEmpty()) {
+                scheduleDetailDao.deleteScheduleDetailsByScheduleId(scheduleDetails[0].scheduleId)
+                scheduleDetails.forEach {
+                    scheduleDetailDao.insert(it)
+                }
             }
         }
     }
